Refactored DatabaseServiceImpl by two databaseService quotes and list
