+++ app/src/main/java/com/softyorch/famousquotes/data/network/databaseService/DatabaseListServiceImpl.kt
@@ -5,7 +5,6 @@ import android.content.Context
 import android.provider.Settings
 import android.provider.Settings.Secure.ANDROID_ID
 import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.Source
 import com.google.firebase.firestore.snapshots
 import com.softyorch.famousquotes.core.FIREBASE_TIMEOUT
 import com.softyorch.famousquotes.core.InternetConnection
@@ -14,14 +13,11 @@ import com.softyorch.famousquotes.data.network.databaseService.typeList.QuoteEdi
 import com.softyorch.famousquotes.data.network.databaseService.typeList.QuoteEditableValuesTypeList
 import com.softyorch.famousquotes.data.network.databaseService.typeList.UserEditableValuesTypeList
 import com.softyorch.famousquotes.data.network.databaseService.utils.tryCatchFireStore
-import com.softyorch.famousquotes.data.network.databaseService.utils.writeLogServiceError
 import com.softyorch.famousquotes.data.network.dto.FavoritesDataDTO
 import com.softyorch.famousquotes.data.network.dto.LikesDataDTO
 import com.softyorch.famousquotes.data.network.response.LikesQuoteResponse
 import com.softyorch.famousquotes.data.network.response.QuoteResponse
-import com.softyorch.famousquotes.domain.interfaces.IDatabaseService
-import com.softyorch.famousquotes.domain.utils.generateRandomId
-import com.softyorch.famousquotes.utils.writeLog
+import com.softyorch.famousquotes.domain.interfaces.IDatabaseListService
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -29,32 +25,22 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.tasks.await
 import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeoutOrNull
 import javax.inject.Inject
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
 
-class DatabaseServiceImpl @Inject constructor(
+class DatabaseListServiceImpl @Inject constructor(
     private val firestore: FirebaseFirestore,
     private val auxFirebaseLists: IAuxFireStoreLists,
     private val internetConnection: InternetConnection,
     private val dispatcher: CoroutineDispatcher = Dispatchers.Default,
     @ApplicationContext private val context: Context,
-) : IDatabaseService {
+): IDatabaseListService {
 
     //Provisional
     @SuppressLint("HardwareIds")
     private val userId = Settings.Secure.getString(context.contentResolver, ANDROID_ID)
-
-    override suspend fun getQuote(id: String): QuoteResponse? = genericGetQuote(id)
-
-    override suspend fun getRandomQuote(): QuoteResponse? = genericGetQuote(null)
-
-    // OJO CON USAR ESTO!!!!!!!!!
-    override suspend fun getAllQuotes(): List<QuoteResponse?>? = genericGetAllQuotes()
 
     override suspend fun setQuoteLikes(updateLikes: LikesDataDTO) {
         val id = updateLikes.id
@@ -151,55 +137,5 @@ class DatabaseServiceImpl @Inject constructor(
             typeList = QuoteEditableQuantityValuesTypeList.Shown(),
             msgError = "Error getting quotes shown"
         )
-
-    /***************************************************************************************/
-    /*********************************** PRIVATE METHODS ***********************************/
-    /***************************************************************************************/
-
-    private suspend fun genericGetQuote(id: String? = null): QuoteResponse? =
-        withTimeoutOrNull(FIREBASE_TIMEOUT) {
-            suspendCancellableCoroutine { cancelableCoroutine ->
-                tryCatchFireStore {
-                    val documentId = id ?: generateRandomId().also {
-                        writeLog(text = "randomId: $it")
-                    }
-
-                    val document = firestore.collection(COLLECTION).document(documentId)
-                    document[Source.CACHE].addOnSuccessListener { snapshot ->
-                        cancelableCoroutine.resume(snapshot.toObject(QuoteResponse::class.java))
-                    }.addOnFailureListener {
-                        document.get().addOnSuccessListener { snapshot ->
-                            cancelableCoroutine.resume(snapshot.toObject(QuoteResponse::class.java))
-                        }.addOnFailureListener { ex ->
-                            writeLogServiceError("Error getting quote", ex)
-                            cancelableCoroutine.resumeWithException(ex)
-                        }
-                    }
-                } ?: cancelableCoroutine.resumeWithException(
-                    auxFirebaseLists.throwService("Error getting quote")
-                )
-            }
-        }
-
-    private suspend fun genericGetAllQuotes(): List<QuoteResponse?>? =
-        withTimeoutOrNull(FIREBASE_TIMEOUT) {
-            suspendCancellableCoroutine { cancelableCoroutine ->
-                tryCatchFireStore {
-                    val document = firestore.collection(COLLECTION)
-                    document[Source.CACHE].addOnSuccessListener { snapshot ->
-                        cancelableCoroutine.resume(snapshot.toObjects(QuoteResponse::class.java))
-                    }.addOnFailureListener {
-                        document.get().addOnSuccessListener { snapshot ->
-                            cancelableCoroutine.resume(snapshot.toObjects(QuoteResponse::class.java))
-                        }.addOnFailureListener { ex ->
-                            writeLogServiceError("Error getting all quotes", ex)
-                            cancelableCoroutine.resumeWithException(ex)
-                        }
-                    }
-                } ?: cancelableCoroutine.resumeWithException(
-                    auxFirebaseLists.throwService("Error getting all quotes")
-                )
-            }
-        }
 
 }
