+++ app/src/main/java/com/softyorch/famousquotes/data/network/databaseService/auxFireStore/IAuxFireStoreLists.kt
@@ -23,6 +23,4 @@ interface IAuxFireStoreLists {
         typeList: QuoteEditableQuantityValuesTypeList,
         msgError: String
     ): MutableList<QuoteResponse>?
-
-    fun throwService(message: String): Throwable
 }
