+++ app/src/main/java/com/softyorch/famousquotes/domain/interfaces/IDatabaseListService.kt
@@ -6,15 +6,12 @@ import com.softyorch.famousquotes.data.network.response.LikesQuoteResponse
 import com.softyorch.famousquotes.data.network.response.QuoteResponse
 import kotlinx.coroutines.flow.Flow
 
-interface IDatabaseService {
-    suspend fun getQuote(id: String): QuoteResponse?
-    suspend fun getRandomQuote(): QuoteResponse?
+interface IDatabaseListService {
     suspend fun setQuoteLikes(updateLikes: LikesDataDTO)
     suspend fun setQuoteShown(id: String)
     suspend fun setQuoteFavorite(updateFavorites: FavoritesDataDTO)
     suspend fun getLikesQuoteFlow(id: String): Flow<LikesQuoteResponse?>
     suspend fun getUserLikeQuote(id: String): Flow<Boolean?>
-    suspend fun getAllQuotes(): List<QuoteResponse?>?
     suspend fun getLikeQuotes(): List<QuoteResponse?>?
     suspend fun getQuotesShown(): List<QuoteResponse?>?
 }
