+++ app/src/main/java/com/softyorch/famousquotes/domain/interfaces/IDatabaseQuoteService.kt
@@ -1,0 +1,9 @@
+package com.softyorch.famousquotes.domain.interfaces
+
+import com.softyorch.famousquotes.data.network.response.QuoteResponse
+
+interface IDatabaseQuoteService {
+    suspend fun getQuote(id: String): QuoteResponse?
+    suspend fun getRandomQuote(): QuoteResponse?
+    suspend fun getAllQuotes(): List<QuoteResponse?>?
+}
