+++ app/src/main/java/com/softyorch/famousquotes/domain/useCases/SetQuoteShown.kt
@@ -1,9 +1,9 @@
 package com.softyorch.famousquotes.domain.useCases
 
-import com.softyorch.famousquotes.domain.interfaces.IDatabaseService
+import com.softyorch.famousquotes.domain.interfaces.IDatabaseListService
 import javax.inject.Inject
 
-class SetQuoteShown @Inject constructor(private val dbService: IDatabaseService) {
+class SetQuoteShown @Inject constructor(private val dbService: IDatabaseListService) {
     suspend operator fun invoke(id: String) {
         dbService.setQuoteShown(id)
     }
