+++ app/src/main/java/com/softyorch/famousquotes/domain/useCases/quoteLikes/GetUserLikeQuote.kt
@@ -1,8 +1,8 @@
 package com.softyorch.famousquotes.domain.useCases.quoteLikes
 
-import com.softyorch.famousquotes.domain.interfaces.IDatabaseService
+import com.softyorch.famousquotes.domain.interfaces.IDatabaseListService
 import javax.inject.Inject
 
-class GetUserLikeQuote @Inject constructor(private val dbService: IDatabaseService) {
+class GetUserLikeQuote @Inject constructor(private val dbService: IDatabaseListService) {
     suspend operator fun invoke(id: String) = dbService.getUserLikeQuote(id)
 }
